+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinDepartment/CgibinDepartmentCreateRequest.cs
@@ -41,6 +41,6 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         /// </summary>
         [Newtonsoft.Json.JsonProperty("order")]
         [System.Text.Json.Serialization.JsonPropertyName("order")]
-        public int? DepartmentOrder { get; set; }
+        public long? DepartmentOrder { get; set; }
     }
 }
