+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinExternalContact/StrategyTag/CgibinExternalContactGetStrategyTagListResponse.cs
@@ -35,7 +35,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
                         /// </summary>
                         [Newtonsoft.Json.JsonProperty("order")]
                         [System.Text.Json.Serialization.JsonPropertyName("order")]
-                        public int Order { get; set; }
+                        public long Order { get; set; }
 
                         /// <summary>
                         /// 获取或设置创建时间戳。
@@ -72,7 +72,7 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
                 /// </summary>
                 [Newtonsoft.Json.JsonProperty("order")]
                 [System.Text.Json.Serialization.JsonPropertyName("order")]
-                public int Order { get; set; }
+                public long Order { get; set; }
 
                 /// <summary>
                 /// 获取或设置创建时间戳。
