+++ src/SKIT.FlurlHttpClient.Wechat.Work/Models/CgibinSchool/Department/CgibinSchoolDepartmentCreateRequest.cs
@@ -78,11 +78,11 @@ namespace SKIT.FlurlHttpClient.Wechat.Work.Models
         public int? RegisterYear { get; set; }
 
         /// <summary>
-        /// 获取或设置排序值。
+        /// 获取或设置次序值。
         /// </summary>
         [Newtonsoft.Json.JsonProperty("order")]
         [System.Text.Json.Serialization.JsonPropertyName("order")]
-        public int? Order { get; set; }
+        public long? Order { get; set; }
 
         /// <summary>
         /// 获取或设置部门管理员列表。
